feat: re-design chapter screen
